pyproject.toml
[project.urls]
homepage = "https://github.com/han-xudong/depth-finger"
issues = "https://github.com/han-xudong/depth-finger/issues"

[project]
name = "depthfinger"
version = "0.1.0"
description = "Depth Finger: A Soft Finger with Depth Sensing for Robotic Manipulation"
authors = [
    {name = "Xudong Han", email = "12231112@mail.sustech.edu.cn"},
    {name = "Fang Wan", email = "wanf@sustech.edu.cn"},
    {name = "Chaoyang Song", email = "songcy@ieee.org"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["robotics", "deep learning", "manipulation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Robotics",
    "Topic :: Software Development :: Build Tools",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "jupyter==1.1.1",
    "matplotlib==3.10.3",
    "numpy==2.2.4",
    "opencv-contrib-python==4.11.0.86",
    "opencv-python==4.11.0.86",
    "protobuf==6.30.1",
    "pyyaml==6.0.2",
    "pyzmq==26.2.0",
    "scipy==1.15.2",
]

[tool.poetry]
requires-poetry = ">=2.1"

[tool.ruff]
line-length = 110
target-version = "py310"
exclude = [
    "tests/data",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "N", "B", "C4", "SIM"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"